refactor(ai): concern-split recruitment planner candidate emit (#3997)
Move gather/evaluate/emit into a part so runRecruitmentPlanner stays well under the ~200-line extract threshold without semantic change.
